+++ src/index.js
@@ -5,6 +5,7 @@ import { config } from './config/index.js';
 import { logger } from './utils/logger.js';
 import { initViewOnceCache } from './core/viewOnceCache.js';
 import { initLockScheduler } from './core/lockScheduler.js';
+import { initGroupGuardian } from './core/groupGuardian.js';
 import { createGroupParticipantsHandler } from './handlers/groupParticipantsHandler.js';
 import { startTelemetry } from './core/telemetry.js';
 import { sendStartupMessage } from './utils/startupMessage.js';
@@ -18,6 +19,7 @@ async function main() {
   await startBaileysClient((sock) => {
     initViewOnceCache(sock);
     initLockScheduler(sock);
+    initGroupGuardian(sock);
     sock.ev.on('messages.upsert', createMessageHandler(sock, commands));
     sock.ev.on('group-participants.update', createGroupParticipantsHandler(sock));
     startTelemetry();